better readme is up
I really wish my pushed commits would show up on the graph and updating the readme is the only way it's letting me actually do it. I do not know how to fix it at this time and I want to keep working on this daily and want people to see I am.